finished most of the code for the forward flux sampling test case (uses /test_case/clean/biphasic_switch.lm as input data). Code still to add: clean program shutdown, final reduce of data collected during fflux sampling (to get the actual reaction constant). Build (-)

+++ src/c/lm/resource/Trajectory.h
@@ -87,6 +87,6 @@ protected:
 }
 }
 
-typedef map<int, lm::resource::Trajectory*> TrajectoryMap;
+typedef map<uint64_t, lm::resource::Trajectory*> TrajectoryMap;
 
 #endif
